% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{def_spatial_domain}
\alias{def_spatial_domain}
\title{defines spatial domain for polygon and points}
\usage{
def_spatial_domain(nc, request)
}
\description{
defines spatial domain for polygon and points
}
