% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{download_with_python}
\alias{download_with_python}
\title{download netcdf from ecmwf using ecmwfapi with python/reticulate. miniconda3 should be installed!}
\usage{
download_with_python(request)
}
\description{
download netcdf from ecmwf using ecmwfapi with python/reticulate. miniconda3 should be installed!
}
