% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{get_nc_spatial_res}
\alias{get_nc_spatial_res}
\title{get ncdf average spatial resolution for this domain}
\usage{
get_nc_spatial_res(request_all)
}
\description{
get ncdf average spatial resolution for this domain
}
