% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{get_reanalysis}
\alias{get_reanalysis}
\title{get reanalysis data from NCEP}
\usage{
get_reanalysis(pt, prefix, fname, varname)
}
\arguments{
\item{pt}{an sf point object in lat long}

\item{lubri}{a lubridate object or a list of two objects defining the time domain or the query}
}
\description{
get reanalysis data from NCEP
}
