% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{lookup_ncep}
\alias{lookup_ncep}
\title{lookup NCEP variable names}
\usage{
lookup_ncep(ncepname)
}
\arguments{
\item{ncepname}{ncep variable name as a string such as 'air','rhum'}
}
\description{
lookup NCEP variable names
}
