% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reanalysis.r
\name{lookup_var}
\alias{lookup_var}
\title{lookup variable names in NCEP and GPCC}
\usage{
lookup_var(var)
}
\arguments{
\item{var}{a meteorological variable name as a string such as 'temperature','relative humidity','u wind','v wind','soil heat flux','net radiation','precipitation rate'}
}
\description{
lookup variable names in NCEP and GPCC
}
