% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping-data.r
\docType{data}
\name{vartable}
\alias{vartable}
\title{Variable names for reanalysis datasets}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.}
\source{
GPCC,NCEP,ERA
}
\usage{
data(vartable)
}
\description{
Variable names for reanalysis datasets
}
\references{
none
}
\keyword{datasets}
